Fix circular dependency by moving decorator (and datahandling import) to chatcommands
